added abstraction for sqlite drivers
