+++ src/driver/websql/WebsqlDriver.ts
@@ -1,112 +1,33 @@
-import {Driver} from "../Driver";
 import {DriverUtils} from "../DriverUtils";
 import {ObjectLiteral} from "../../common/ObjectLiteral";
 import {ColumnMetadata} from "../../metadata/ColumnMetadata";
 import {DriverOptionNotSetError} from "../../error/DriverOptionNotSetError";
-import {DateUtils} from "../../util/DateUtils";
 import {WebsqlQueryRunner} from "./WebsqlQueryRunner";
 import {Connection} from "../../connection/Connection";
-import {RdbmsSchemaBuilder} from "../../schema-builder/RdbmsSchemaBuilder";
 import {WebSqlConnectionOptions} from "./WebSqlConnectionOptions";
-import {MappedColumnTypes} from "../types/MappedColumnTypes";
-import {ColumnType} from "../types/ColumnTypes";
-import {DataTypeDefaults} from "../types/DataTypeDefaults";
-import {ColumnSchema} from "../../schema-builder/schema/ColumnSchema";
-import {RandomGenerator} from "../../util/RandomGenerator";
+import {AbstractSqliteDriver} from "../sqlite-abstract/AbstractSqliteDriver";
 
 /**
  * Organizes communication with WebSQL in the browser.
  */
-export class WebsqlDriver implements Driver {
+export class WebsqlDriver extends AbstractSqliteDriver {
 
     // -------------------------------------------------------------------------
     // Public Properties
     // -------------------------------------------------------------------------
 
     /**
-     * Connection used by driver.
-     */
-    connection: Connection;
-
-    /**
      * Connection options.
      */
     options: WebSqlConnectionOptions;
-
-    /**
-     * Default values of length, precision and scale depends on column data type.
-     * Used in the cases when length/precision/scale is not specified by user.
-     */
-    dataTypeDefaults: DataTypeDefaults;
-
-    // -------------------------------------------------------------------------
-    // Public Implemented Properties
-    // -------------------------------------------------------------------------
-
-    /**
-     * Indicates if tree tables are supported by this driver.
-     */
-    treeSupport = true;
-
-    /**
-     * Gets list of supported column data types by a driver.
-     *
-     * @see https://www.tutorialspoint.com/sqlite/sqlite_data_types.htm
-     * @see https://sqlite.org/datatype3.html
-     */
-    supportedDataTypes: ColumnType[] = [
-        "int",
-        "integer",
-        "tinyint",
-        "smallint",
-        "mediumint",
-        "bigint",
-        "int2",
-        "int8",
-        "integer",
-        "character",
-        "varchar",
-        "varying character",
-        "nchar",
-        "native character",
-        "nvarchar",
-        "text",
-        "clob",
-        "text",
-        "blob",
-        "real",
-        "double",
-        "double precision",
-        "float",
-        "real",
-        "numeric",
-        "decimal",
-        "boolean",
-        "date",
-        "datetime",
-    ];
-
-    /**
-     * Orm has special columns and we need to know what database column types should be for those types.
-     * Column types are driver dependant.
-     */
-    mappedDataTypes: MappedColumnTypes = {
-        createDate: "datetime",
-        createDateDefault: "datetime('now')",
-        updateDate: "datetime",
-        updateDateDefault: "datetime('now')",
-        version: "int",
-        treeLevel: "int",
-        migrationName: "varchar",
-        migrationTimestamp: "bigint",
-    };
 
     // -------------------------------------------------------------------------
     // Constructor
     // -------------------------------------------------------------------------
 
     constructor(connection: Connection) {
-        this.connection = connection;
+        super(connection);
+        
         this.options = connection.options as WebSqlConnectionOptions;
         Object.assign(connection.options, DriverUtils.buildDriverOptions(connection.options)); // todo: do it better way
 
@@ -126,14 +47,8 @@ export class WebsqlDriver implements Driver {
 
     /**
      * Performs connection to the database.
-     * Based on pooling options, it can either create connection immediately,
-     * either create a pool and create connection when needed.
      */
-    connect(): Promise<void> {
-        return Promise.resolve();
-    }
-
-    afterConnect(): Promise<void> {
+    async connect(): Promise<void> {
         return Promise.resolve();
     }
 
@@ -153,17 +68,32 @@ export class WebsqlDriver implements Driver {
     }
 
     /**
-     * Creates a schema builder used to build and sync a schema.
+     * Creates a query runner used to execute database queries.
      */
-    createSchemaBuilder() {
-        return new RdbmsSchemaBuilder(this.connection);
+    createQueryRunner(): WebsqlQueryRunner {
+        return new WebsqlQueryRunner(this);
     }
 
     /**
-     * Creates a query runner used to execute database queries.
+     * Prepares given value to a value to be persisted, based on its column type and metadata.
      */
-    createQueryRunner() {
-        return new WebsqlQueryRunner(this);
+    preparePersistentValue(value: any, columnMetadata: ColumnMetadata): any {
+        if (columnMetadata.type === "json") {
+            return JSON.stringify(value);
+        }
+
+        return super.preparePersistentValue(value, columnMetadata);
+    }
+
+    /**
+     * Prepares given value to a value to be persisted, based on its column type or metadata.
+     */
+    prepareHydratedValue(value: any, columnMetadata: ColumnMetadata): any {
+        if (columnMetadata.type === "json") {
+            return JSON.parse(value);
+        }
+
+        return super.prepareHydratedValue(value, columnMetadata);
     }
 
     /**
@@ -194,138 +124,4 @@ export class WebsqlDriver implements Driver {
     escape(columnName: string): string {
         return columnName; // "`" + columnName + "`";
     }
-
-    /**
-     * Prepares given value to a value to be persisted, based on its column type and metadata.
-     */
-    preparePersistentValue(value: any, columnMetadata: ColumnMetadata): any {
-        if (value === null || value === undefined)
-            return value;
-
-        if (columnMetadata.type === Boolean) {
-            return value === true ? 1 : 0;
-
-        } else if (columnMetadata.type === "date") {
-            return DateUtils.mixedDateToDateString(value);
-
-        } else if (columnMetadata.type === "time") {
-            return DateUtils.mixedDateToTimeString(value);
-
-        } else if (columnMetadata.type === "datetime") {
-            return DateUtils.mixedDateToUtcDatetimeString(value);
-
-        } else if (columnMetadata.isGenerated && columnMetadata.generationStrategy === "uuid" && !value) {
-            return RandomGenerator.uuid4();
-
-        } else if (columnMetadata.type === "json") {
-            return JSON.stringify(value);
-
-        } else if (columnMetadata.type === "simple-array") {
-            return DateUtils.simpleArrayToString(value);
-        }
-
-        return value;
-    }
-
-    /**
-     * Prepares given value to a value to be persisted, based on its column type or metadata.
-     */
-    prepareHydratedValue(value: any, columnMetadata: ColumnMetadata): any {
-        if (value === null || value === undefined)
-            return value;
-            
-        if (columnMetadata.type === Boolean) {
-            return value ? true : false;
-
-        } else if (columnMetadata.type === "datetime") {
-            return DateUtils.normalizeHydratedDate(value);
-
-        } else if (columnMetadata.type === "date") {
-            return DateUtils.mixedDateToDateString(value);
-
-        } else if (columnMetadata.type === "time") {
-            return DateUtils.mixedTimeToString(value);
-
-        } else if (columnMetadata.type === "json") {
-            return JSON.parse(value);
-
-        } else if (columnMetadata.type === "simple-array") {
-            return DateUtils.stringToSimpleArray(value);
-        }
-
-        return value;
-    }
-
-    /**
-     * Creates a database type from a given column metadata.
-     */
-    normalizeType(column: { type?: ColumnType, length?: number, precision?: number, scale?: number }): string {
-        if (column.type === Number || column.type === "int") {
-            return "integer";
-
-        } else if (column.type === String) {
-            return "varchar";
-
-        } else if (column.type === Date) {
-            return "datetime";
-
-        } else if ((column.type as any) === Buffer) {
-            return "blob";
-
-        } else if (column.type === Boolean) {
-            return "boolean";
-
-        } else if (column.type === "uuid") {
-            return "varchar";
-
-        } else if (column.type === "simple-array") {
-            return "text";
-
-        } else {
-            return column.type as string || "";
-        }
-    }
-
-    /**
-     * Normalizes "default" value of the column.
-     */
-    normalizeDefault(column: ColumnMetadata): string {
-        if (typeof column.default === "number") {
-            return "" + column.default;
-
-        } else if (typeof column.default === "boolean") {
-            return column.default === true ? "1" : "0";
-
-        } else if (typeof column.default === "function") {
-            return column.default();
-
-        } else if (typeof column.default === "string") {
-            return `'${column.default}'`;
-
-        } else {
-            return column.default;
-        }
-    }
-
-    createFullType(column: ColumnSchema): string {
-        let type = column.type;
-
-        if (column.length) {
-            type += "(" + column.length + ")";
-        } else if (column.precision && column.scale) {
-            type += "(" + column.precision + "," + column.scale + ")";
-        } else if (column.precision) {
-            type +=  "(" + column.precision + ")";
-        } else if (column.scale) {
-            type +=  "(" + column.scale + ")";
-        } else  if (this.dataTypeDefaults && this.dataTypeDefaults[column.type] && this.dataTypeDefaults[column.type].length) {
-            type +=  "(" + this.dataTypeDefaults[column.type].length + ")";
-        }
-
-        if (column.isArray)
-            type += " array";
-
-        return type;
-    }
-
 }
